% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder.R
\name{outerwidthofcylinder}
\alias{outerwidthofcylinder}
\title{cylinder outer width}
\usage{
outerwidthofcylinder(r, h)
}
\description{
This function computes the outer width of the cylinder.
}
\examples{

outerwidthofcylinder(3, 5)
}
